Relocate Comment & Updown
Put the comment and updown models under the "content" app. Add admin sections for them.
