+++ backend-app/interesthub/content/admin.py
@@ -1,5 +1,5 @@
 from django.contrib import admin
-from .models import Content, ContentType
+from .models import *
 from components.models import Component
 
 
@@ -14,5 +14,15 @@ class ContentTypeAdmin(admin.ModelAdmin):
     readonly_fields = ("id", "created_date", "modified_date",)
     list_display = ["id", "name"]
 
+class CommentAdmin(admin.ModelAdmin):
+    fields = ["text", "owner", "content"]
+    list_display = ["id", "owner", "content"]
+
+class UpDownAdmin(admin.ModelAdmin):
+    fields = ["isUp", "owner", "content"]
+    list_display = ["id", "isUp", "owner", "content"]
+
 admin.site.register(ContentType, ContentTypeAdmin)
 admin.site.register(Content, ContentAdmin)
+admin.site.register(Comment, CommentAdmin)
+admin.site.register(UpDown, UpDownAdmin)
