+++ backend-app/interesthub/interesthub/settings.py
@@ -46,8 +46,6 @@ INSTALLED_APPS = [
     'group',
     'content',
     'components',
-    'comment',
-    'upDown',
     'user',
 ]
