+++ backend-app/interesthub/interesthub/urls.py
@@ -18,11 +18,9 @@ from django.conf.urls import url, include
 from rest_framework import routers
 from dummy import views
 from group.views import GroupContentList
-from content.views import ContentViewSet, ContentTypeViewSet
+from content.views import *
 from components.views import ComponentViewSet
 from group.views import *
-from comment.views import *
-from upDown.views import *
 from user.views import *
 from rest_framework.urlpatterns import format_suffix_patterns
 from rest_framework_jwt.views import obtain_jwt_token
