--- backend-app/interesthub/upDown/apps.py
@@ -1,5 +1,0 @@
-from django.apps import AppConfig
-
-
-class UpdownConfig(AppConfig):
-    name = 'upDown'
